Show details of a game log entry, refactored CardManager to keep a database of cards and build default decks based on card ID

+++ GameThing/UI/FadingTextPanel.cs
@@ -51,7 +51,7 @@ namespace GameThing.UI
 			}
 		}
 
-		public override void Draw(SpriteBatch spriteBatch, int x, int y)
+		public override void Draw(SpriteBatch spriteBatch, float x, float y)
 		{
 			if (!showing)
 				return;
